PHPStan fixes

+++ app/helpers.php
@@ -51,7 +51,7 @@ if (! function_exists('yesno')) {
      * return "yes" or "no" based on condition
      *
      * @param $condition
-     * @return void
+     * @return string
      */
     function yesno($condition): string
     {
@@ -70,7 +70,7 @@ if (! function_exists('human_filesize')) {
     function human_filesize($bytes, $dec = 2): string {
 
         $size = ['B', 'kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-        $factor = floor((strlen($bytes) - 1) / 3);
+        $factor = (int) floor((strlen($bytes) - 1) / 3);
         if ($factor == 0) $dec = 0;
 
         return sprintf("%.{$dec}f %s", $bytes / (1024 ** $factor), $size[$factor]);
